refactor - 결국 ClassScanner 클래스 분리하는걸로 결정
