+++ src/main/java/util/RequestMapper.java
@@ -24,7 +24,7 @@ public class RequestMapper {
     public static final Map<String, Method> REQUEST_MAP = new HashMap<>();
 
     static {
-        List<Class<?>> controllers = scanControllers("controller");
+        List<Class<?>> controllers = ClassScanner.scanControllers("controller");
 
         Class<? extends Annotation> requestMapping = RequestMapping.class;
         for (Class<?> c : controllers) {
@@ -37,59 +37,6 @@ public class RequestMapper {
                 }
             }
         }
-    }
-
-    public static List<Class<?>> scanControllers(String basePackage) {
-        List<Class<?>> controllerClasses = new ArrayList<>();
-
-        String basePath = basePackage.replace(".", File.separator);
-        String packagePath = "src/main/java/" + basePath;
-
-        List<File> classFiles = listFiles(packagePath);
-
-        for (File file : classFiles) {
-            String className = getClassName(file, basePath);
-            try {
-                Class<?> clazz = Class.forName(className);
-                if (hasControllerAnnotation(clazz)) controllerClasses.add(clazz);
-            } catch (ClassNotFoundException e) {
-                logger.error(e.getMessage());
-            }
-        }
-
-        return controllerClasses;
-    }
-
-    private static List<File> listFiles(String directoryPath) {
-        List<File> classFiles = new ArrayList<>();
-        File directory = new File(directoryPath);
-
-        if (directory.exists() && directory.isDirectory()) {
-            File[] files = directory.listFiles();
-            if (files != null) {
-                for (File file : files) {
-                    if (file.isFile() && (file.getName().endsWith(".java") || file.getName().endsWith(".class")))
-                        classFiles.add(file);
-                }
-            }
-        }
-
-        return classFiles;
-    }
-
-    private static String getClassName(File file, String basePath) {
-        String fileName = file.getAbsolutePath().replace(File.separator, ".");
-        int startIndex = fileName.indexOf(basePath);
-        int endIndex = fileName.lastIndexOf(".");
-        return fileName.substring(startIndex, endIndex);
-    }
-
-    private static boolean hasControllerAnnotation(Class<?> clazz) {
-        Annotation[] annotations = clazz.getDeclaredAnnotations();
-        for (Annotation annotation : annotations) {
-            if (annotation.annotationType().getSimpleName().equals("Controller")) return true;
-        }
-        return false;
     }
 
     public static Method getMethod(HttpRequest request) {
